+++ changelog_entry.yaml
@@ -1,4 +1,4 @@
 - bump: minor
   changes:
     added:
-    - Sk Disability amount tax credit.
+    - Saskatchewan disability amount.
