--- policyengine_canada/parameters/gov/provinces/sk/tax/income/credits/sk_disability_amount/amount.yaml
@@ -1,13 +1,0 @@
-description: Saskatchewan provides this tax-credit amount for the qualified disabled.
-values:
-  2023-01-01: 10_405
-metadata:
-  unit: currency-CAD
-  label: Saskatchewan disability tax credit amount
-  reference:
-  - title: 2023 Saskatchewan TD1SK Personal Tax Credits Return
-    href: https://www.canada.ca/content/dam/cra-arc/formspubs/pbg/td1sk/td1sk-23e.pdf
-  - title: 2023 Saskatchewan TD1SK-WS Personal Tax Credits Return
-    href: https://www.canada.ca/content/dam/cra-arc/formspubs/pbg/td1sk-ws/td1sk-ws-23e.pdf
-  - title: Saskatchewan I2-01 The Income Tax Act, 2000
-    href: https://pubsaskdev.blob.core.windows.net/pubsask-prod/806/I2-01.pdf
