+++ policyengine_canada/tests/gov/provinces/sk/tax/income/credits/sk_disability_amount.yaml
@@ -1,17 +1,35 @@
-- name:  Have Disability Tax Credit Certificate
-  period: 2024
+- name: Head and disabled
+  period: 2023
   input:
     province_code: SK
-    is_disable_certificate: True
-
+    is_head: true
+    is_disabled: true
   output:
     sk_disability_amount: 10_405
 
-- name:  Do Not Have Disability Tax Credit Certificate
+- name: Head, not disabled
   period: 2024
   input:
     province_code: SK
-    is_disable_certificate: False
+    is_head: true
+    is_disabled: false
+  output:
+    sk_disability_amount: 0
 
+- name: Head and disabled 2022
+  period: 2022
+  input:
+    province_code: SK
+    is_head: true
+    is_disabled: true
+  output:
+    sk_disability_amount: 9_789
+
+- name: Disabled, not head
+  period: 2024
+  input:
+    province_code: SK
+    is_head: false
+    is_disabled: true
   output:
     sk_disability_amount: 0
